+++ lib/types.d.ts
@@ -38,18 +38,18 @@ export declare type RolesAt = {
 };
 export interface AccessClaims {
     name?: string;
-    user?: string;
+    user: string;
     client?: string;
     roles: RolesAt;
     scopes: Scopes[];
 }
 export interface RefreshClaims {
-    owner: string;
+    user: string;
     grant_type: GrantTypes;
 }
 export interface Claims extends AccessClaims, RefreshClaims {
-    exp: Date;
-    iat: Date;
+    exp: number;
+    iat: number;
 }
 export interface ResourceAttributeAsMapConstructor extends Array<Resource | string[]> {
     0: Resource;